requirements.txt
playwright
google-generativeai
python-dotenv
pydantic
jsonlines
